requirements.txt
asgiref==3.8.1
boto3==1.35.76
botocore==1.35.76
crispy-bootstrap5==2024.10
dj-database-url==0.5.0
Django==4.2.16
django-allauth==65.2.0
django-crispy-forms==2.3
django-storages==1.14.4
gunicorn==23.0.0
jmespath==1.0.1
pillow==11.0.0
psycopg2==2.9.10
s3transfer==0.10.4
sqlparse==0.5.2
stripe==11.2.0
whitenoise==6.8.2
